+++ MostActiveFriend/FriendWall.cs
@@ -8,7 +8,7 @@ namespace MostActiveFriend
     public class FriendWall
     {
         public List<Post> postList { get; set; }
-        public int user_id { get; set; }
+        public long user_id { get; set; }
         public int count { get; set; }
 
         public FriendWall()
